Add notifications in the app. Allow all kinds of foreground, background or terminated state notifications. Store FCM tokens in the database when login.

+++ lib/view_model/after_auth_view_models/profile_view_models/profile_page_view_model.dart
@@ -1,6 +1,7 @@
 import 'package:currency_picker/currency_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:graphql_flutter/graphql_flutter.dart';
 import 'package:hive/hive.dart';
 import 'package:qr_flutter/qr_flutter.dart';
 import 'package:social_share/social_share.dart';
@@ -51,26 +52,34 @@ class ProfilePageViewModel extends BaseModel {
         reverse: true,
         dialogSubTitle: 'Are you sure you want to logout?',
         successText: 'Logout',
-        success: () {
-          navigationService.pop();
-          navigationService.pushDialog(
-            const CustomProgressDialog(
-              key: Key('LogoutProgress'),
-            ),
-          );
-          Future.delayed(const Duration(seconds: 1)).then((value) {
-            user = Hive.box<User>('currentUser');
-            url = Hive.box('url');
-            organisation = Hive.box<OrgInfo>('currentOrg');
-            user.clear();
-            url.clear();
-            organisation.clear();
-            navigationService.removeAllAndPush(
-              '/selectLang',
-              '/',
-              arguments: '0',
-            );
-          });
+        success: () async {
+          try {
+            final result = await databaseFunctions
+                .gqlAuthMutation(queries.logout()) as QueryResult?;
+            if (result != null && result.data!['logout'] == true) {
+              navigationService.pop();
+              navigationService.pushDialog(
+                const CustomProgressDialog(
+                  key: Key('LogoutProgress'),
+                ),
+              );
+              Future.delayed(const Duration(seconds: 1)).then((value) {
+                user = Hive.box<User>('currentUser');
+                url = Hive.box('url');
+                organisation = Hive.box<OrgInfo>('currentOrg');
+                user.clear();
+                url.clear();
+                organisation.clear();
+                navigationService.removeAllAndPush(
+                  '/selectLang',
+                  '/',
+                  arguments: '0',
+                );
+              });
+            }
+          } catch (e) {
+            print(e);
+          }
         },
       ),
     );
